feat: added imported contacts to new chat page
